--- src/components/TaskStatusIcon/CannotRun.tsx
@@ -1,10 +1,0 @@
-export const CannotRun: React.FC = () => (
-  <svg width="24" height="24" viewBox="0 0 24 24" fill="none">
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M24 12C24 18.6274 18.6274 24 12 24C5.37258 24 0 18.6274 0 12C0 5.37258 5.37258 0 12 0C18.6274 0 24 5.37258 24 12ZM19.5 12C19.5 16.1421 16.1421 19.5 12 19.5C10.7483 19.5 9.56827 19.1934 8.53084 18.6512L18.6512 8.53082C19.1934 9.56826 19.5 10.7483 19.5 12ZM5.34885 15.4692L15.4692 5.34884C14.4317 4.80662 13.2517 4.5 12 4.5C7.85786 4.5 4.5 7.85786 4.5 12C4.5 13.2517 4.80662 14.4317 5.34885 15.4692Z"
-      fill="#21313C"
-    />
-  </svg>
-);
