--- src/components/TaskStatusIcon/Failed.tsx
@@ -1,8 +1,0 @@
-export const Failed: React.FC = () => (
-  <svg width="20" height="20" viewBox="0 0 20 20" fill="none">
-    <path
-      d="M17.7783 0.807588C16.9972 0.0265398 15.7309 0.0265394 14.9498 0.807588L10.0001 5.75734L5.05035 0.807588C4.2693 0.0265396 3.00297 0.0265398 2.22192 0.807588L0.80771 2.2218C0.0266619 3.00285 0.0266615 4.26918 0.80771 5.05023L5.75746 9.99998L0.80771 14.9497C0.0266626 15.7308 0.0266619 16.9971 0.80771 17.7782L2.22192 19.1924C3.00297 19.9734 4.2693 19.9734 5.05035 19.1924L10.0001 14.2426L14.9498 19.1924C15.7309 19.9734 16.9972 19.9734 17.7783 19.1924L19.1925 17.7782C19.9735 16.9971 19.9735 15.7308 19.1925 14.9497L14.2427 9.99998L19.1925 5.05023C19.9735 4.26918 19.9735 3.00285 19.1925 2.2218L17.7783 0.807588Z"
-      fill="#CF4A22"
-    />
-  </svg>
-);
