--- src/components/TaskStatusIcon/Running.tsx
@@ -1,12 +1,0 @@
-export const Running: React.FC = () => (
-  <svg width="30" height="24" viewBox="0 0 30 24" fill="none">
-    <path
-      d="M15.0658 0C20.4637 0 24.9663 3.43552 26.0001 8L28.8287 8C29.911 8 30.4121 9.34404 29.594 10.0526L24.6307 14.3509C24.1914 14.7313 23.5394 14.7313 23.1002 14.3509L18.1369 10.0526C17.3187 9.34404 17.8198 8 18.9021 8H21.7954C20.8579 5.82912 18.4344 4 15.0658 4C13.0132 4 11.3116 4.67908 10.0831 5.70078C9.59719 6.10496 8.84858 6.10539 8.45099 5.61403L7.22998 4.10508C6.91086 3.7107 6.9345 3.13584 7.31819 2.80394C9.3251 1.06789 12.0563 0 15.0658 0Z"
-      fill="#86681D"
-    />
-    <path
-      d="M6.89982 9.6853C6.46057 9.30489 5.80858 9.30489 5.36932 9.6853L0.406033 13.9836C-0.412112 14.6922 0.0889775 16.0362 1.17128 16.0362H4.00821C5.05755 20.5826 9.55061 24 14.9342 24C17.9437 24 20.6749 22.9321 22.6818 21.1961C23.0655 20.8642 23.0891 20.2893 22.77 19.8949L21.549 18.386C21.1514 17.8946 20.4028 17.895 19.9169 18.2992C18.6884 19.3209 16.9868 20 14.9342 20C11.5844 20 9.16906 18.1912 8.22039 16.0362H11.0979C12.1802 16.0362 12.6813 14.6922 11.8631 13.9836L6.89982 9.6853Z"
-      fill="#86681D"
-    />
-  </svg>
-);
