--- src/components/TaskStatusIcon/Success.tsx
@@ -1,10 +1,0 @@
-export const Success: React.FC = () => (
-  <svg width="25" height="21" viewBox="0 0 25 21" fill="none">
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M8.61171 12.1008L19.5223 1.19022C20.3033 0.409169 21.5697 0.409169 22.3507 1.19022L23.7649 2.60443C24.546 3.38548 24.546 4.65181 23.7649 5.43286L9.63056 19.5672C8.77644 20.4213 7.36591 20.3292 6.63017 19.3712L0.532819 11.4321C-0.139981 10.5561 0.0247636 9.30053 0.900786 8.62773L2.09042 7.71407C3.18545 6.87307 4.75491 7.07901 5.59591 8.17403L8.61171 12.1008Z"
-      fill="#13AA52"
-    />
-  </svg>
-);
