--- src/components/TaskStatusIcon/SystemFailure.tsx
@@ -1,8 +1,0 @@
-export const SystemFailure: React.FC = () => (
-  <svg width="29" height="28" viewBox="0 0 29 28" fill="none">
-    <path
-      d="M14.76 19.56C14.5979 19.3931 14.4507 19.2124 14.32 19.02C13.6892 19.0018 13.0664 18.8732 12.48 18.64C11.788 18.4092 11.1507 18.039 10.6073 17.5524C10.0639 17.0657 9.62599 16.4729 9.32053 15.8105C9.01507 15.1481 8.84857 14.4301 8.83129 13.7009C8.81401 12.9716 8.94633 12.2466 9.22007 11.5705C9.4938 10.8943 9.90317 10.2814 10.4229 9.76961C10.9426 9.25778 11.5617 8.85786 12.242 8.59452C12.9222 8.33118 13.6492 8.21 14.3781 8.23846C15.107 8.26692 15.8223 8.44441 16.48 8.75998C17.3347 9.10972 18.0822 9.67829 18.6474 10.4085C19.2126 11.1388 19.5757 12.0049 19.7 12.92C20.0518 13.0981 20.3753 13.3272 20.66 13.6L22.42 15.36L24.2 13.6C24.7376 13.0545 25.441 12.7028 26.2 12.6H26.44C26.44 12.36 26.44 12.1 26.44 11.86L27.42 11.14C27.6835 10.9457 27.8741 10.6686 27.9611 10.353C28.0482 10.0375 28.0266 9.70184 27.9 9.39998L27.1 7.19998C26.9715 6.89766 26.7474 6.64587 26.4619 6.48328C26.1765 6.32069 25.8456 6.2563 25.52 6.29998L24.32 6.49998C23.5936 5.49779 22.7166 4.614 21.72 3.87998L21.9 2.67998C21.9425 2.35717 21.8773 2.02936 21.7146 1.74733C21.5519 1.4653 21.3007 1.2448 21 1.11998L18.82 0.199984C18.5074 0.060255 18.1564 0.0316753 17.8253 0.118993C17.4943 0.20631 17.203 0.404269 17 0.679984L16.28 1.67998C15.0615 1.47998 13.8185 1.47998 12.6 1.67998L11.86 0.699984C11.6691 0.438117 11.3955 0.248284 11.0834 0.16118C10.7713 0.0740766 10.4389 0.0947946 10.14 0.219984L7.94 1.11998C7.64089 1.24534 7.39203 1.46672 7.23268 1.74919C7.07334 2.03166 7.01258 2.35915 7.06 2.67998L7.32 3.99998C6.32815 4.73171 5.45173 5.60814 4.72 6.59998L3.5 6.41998C3.17954 6.37604 2.8535 6.43829 2.57177 6.59721C2.29004 6.75614 2.06813 7.00298 1.94 7.29998L1 9.39998C0.873367 9.70184 0.851848 10.0375 0.938897 10.353C1.02595 10.6686 1.21651 10.9457 1.48 11.14L2.46 11.86C2.27079 13.0794 2.27079 14.3206 2.46 15.54L1.48 16.28C1.21813 16.4709 1.0283 16.7445 0.941197 17.0566C0.854093 17.3687 0.874811 17.7011 1 18L1.9 20.2C2.02617 20.5029 2.25069 20.7544 2.53734 20.9141C2.82399 21.0737 3.15606 21.1322 3.48 21.08L4.68 20.9C5.41173 21.8918 6.28815 22.7683 7.28 23.5L7.1 24.72C7.06092 25.0424 7.12753 25.3688 7.28982 25.6501C7.45211 25.9314 7.70133 26.1524 8 26.28L10.18 27.18C10.4819 27.3066 10.8175 27.3281 11.1331 27.2411C11.4486 27.154 11.7258 26.9635 11.92 26.7L12.64 25.72C13.0046 25.7731 13.3718 25.8065 13.74 25.82C13.7002 25.323 13.767 24.8232 13.936 24.3541C14.105 23.885 14.3723 23.4574 14.72 23.1L16.5 21.34L14.76 19.56Z"
-      fill="#36367F"
-    />
-  </svg>
-);
