--- src/components/TaskStatusIcon/TimedOut.tsx
@@ -1,12 +1,0 @@
-export const TimedOut: React.FC = () => (
-  <svg width="30" height="29" viewBox="0 0 30 29" fill="none">
-    <path
-      d="M25.68 23L28.7 19.98C28.8149 19.8682 28.9062 19.7345 28.9686 19.5867C29.031 19.439 29.0631 19.2803 29.0631 19.12C29.0631 18.9597 29.031 18.8009 28.9686 18.6532C28.9062 18.5055 28.8149 18.3718 28.7 18.26L27.84 17.4L27.42 17.12H27.28C27.0724 17.0668 26.8546 17.0684 26.6478 17.1245C26.441 17.1805 26.2523 17.2892 26.1 17.44L23 20.44L20.06 17.42C19.9482 17.3051 19.8145 17.2138 19.6667 17.1514C19.519 17.089 19.3603 17.0569 19.2 17.0569C19.0397 17.0569 18.8809 17.089 18.7332 17.1514C18.5855 17.2138 18.4518 17.3051 18.34 17.42L17.48 18.28C17.3651 18.3918 17.2738 18.5255 17.2114 18.6732C17.149 18.8209 17.1169 18.9797 17.1169 19.14C17.1169 19.3003 17.149 19.459 17.2114 19.6067C17.2738 19.7545 17.3651 19.8882 17.48 20L20.5 23L17.48 26.04C17.3392 26.1821 17.2361 26.3572 17.1803 26.5493C17.1244 26.7415 17.1174 26.9445 17.16 27.14V27.3L17.42 27.76L18.28 28.64C18.511 28.8624 18.8193 28.9867 19.14 28.9867C19.4607 28.9867 19.7689 28.8624 20 28.64L23 25.62L26.02 28.64C26.2544 28.8637 26.566 28.9886 26.89 28.9886C27.214 28.9886 27.5256 28.8637 27.76 28.64L28.62 27.76C28.7349 27.6482 28.8262 27.5145 28.8886 27.3667C28.951 27.219 28.9831 27.0603 28.9831 26.9C28.9831 26.7397 28.951 26.5809 28.8886 26.4332C28.8262 26.2855 28.7349 26.1518 28.62 26.04L25.68 23Z"
-      fill="#CF4A22"
-    />
-    <path
-      d="M14.26 0.999991C12.4687 0.93054 10.6818 1.22339 9.00637 1.86099C7.33094 2.49858 5.80145 3.46782 4.50954 4.71061C3.21762 5.95341 2.18986 7.4442 1.48783 9.09367C0.785804 10.7431 0.42395 12.5173 0.42395 14.31C0.42395 16.1026 0.785804 17.8768 1.48783 19.5263C2.18986 21.1758 3.21762 22.6666 4.50954 23.9094C5.80145 25.1522 7.33094 26.1214 9.00637 26.759C10.6818 27.3966 12.4687 27.6894 14.26 27.62H15.06H15C14.8792 27.0754 14.8984 26.5092 15.0557 25.974C15.2131 25.4389 15.5036 24.9524 15.9 24.56L17.44 23L15.96 21.5C15.6525 21.1892 15.4093 20.8207 15.2446 20.4157C15.0799 20.0107 14.9967 19.5772 15 19.14C15.0022 18.2391 15.3619 17.3759 16 16.74L16.86 15.88C17.4955 15.293 18.3351 14.9772 19.2 15C19.6448 14.9974 20.0858 15.0826 20.4977 15.2508C20.9095 15.4189 21.2841 15.6667 21.6 15.98L23.08 17.48L24.58 15.98C24.8962 15.6616 25.2737 15.4106 25.6896 15.2421C26.1056 15.0737 26.5513 14.9913 27 15C27.1928 14.9793 27.3872 14.9793 27.58 15C27.58 14.78 27.58 14.58 27.58 14.36C27.5853 12.6074 27.2446 10.871 26.5776 9.25036C25.9105 7.62969 24.9302 6.15658 23.6928 4.91546C22.4554 3.67434 20.9853 2.68962 19.3666 2.01773C17.7479 1.34584 16.0126 0.999983 14.26 0.999991ZM9.59999 15.7C9.42271 15.7 9.24716 15.6651 9.08337 15.5972C8.91958 15.5294 8.77076 15.4299 8.6454 15.3046C8.52004 15.1792 8.4206 15.0304 8.35276 14.8666C8.28491 14.7028 8.24999 14.5273 8.24999 14.35C8.24999 14.1727 8.28491 13.9972 8.35276 13.8334C8.4206 13.6696 8.52004 13.5208 8.6454 13.3954C8.77076 13.27 8.91958 13.1706 9.08337 13.1028C9.24716 13.0349 9.42271 13 9.59999 13H13.6V7.19999C13.5995 7.02052 13.6361 6.84289 13.7076 6.67827C13.7791 6.51366 13.8839 6.36563 14.0154 6.24351C14.1469 6.12139 14.3023 6.02782 14.4717 5.9687C14.6412 5.90958 14.8211 5.88618 15 5.89999C15.3466 5.89995 15.6794 6.03627 15.9264 6.27951C16.1734 6.52275 16.3147 6.85338 16.32 7.19999V14.38C16.32 14.7301 16.1809 15.0658 15.9334 15.3134C15.6858 15.5609 15.3501 15.7 15 15.7H9.59999Z"
-      fill="#CF4A22"
-    />
-  </svg>
-);
