--- src/components/TaskStatusIcon/WillNotRun.tsx
@@ -1,20 +1,0 @@
-export const WillNotRun: React.FC = () => (
-  <svg width="30" height="28" viewBox="0 0 30 28" fill="none">
-    <path
-      d="M7.64 6.54C8.16669 6.48503 8.66672 6.28078 9.08126 5.95127C9.49581 5.62176 9.80762 5.18071 9.98 4.68V2.82C9.80863 2.31684 9.49758 1.87286 9.08321 1.53995C8.66884 1.20703 8.16826 0.998944 7.64 0.940002C7.39405 0.939988 7.15054 0.988753 6.92356 1.08348C6.69658 1.1782 6.49064 1.31699 6.31766 1.49184C6.14468 1.66668 6.00809 1.87409 5.9158 2.10207C5.82351 2.33005 5.77736 2.57407 5.78 2.82V4.68C5.77732 4.925 5.8236 5.16807 5.91613 5.39495C6.00866 5.62182 6.14556 5.82794 6.31881 6.00119C6.49207 6.17444 6.69818 6.31135 6.92505 6.40387C7.15193 6.4964 7.395 6.54268 7.64 6.54Z"
-      fill="#5D6C74"
-    />
-    <path
-      d="M17.56 20.62L15.84 18.9C15.2394 18.299 14.8809 17.4979 14.833 16.6496C14.785 15.8012 15.0509 14.9649 15.58 14.3V12.72H17.18C17.7555 12.2829 18.4573 12.0443 19.18 12.04C19.6344 12.0373 20.0846 12.1262 20.5039 12.3015C20.9231 12.4767 21.3027 12.7347 21.62 13.06L23.36 14.78L25.36 12.78L25.56 12.66C25.934 12.4091 26.3459 12.2201 26.78 12.1V8.10001C26.7747 7.09911 26.3748 6.1407 25.667 5.43295C24.9593 4.72519 24.0009 4.32526 23 4.32C23.0378 4.84087 22.9679 5.36394 22.7946 5.85658C22.6213 6.34922 22.3483 6.80085 21.9927 7.18327C21.6371 7.5657 21.2064 7.87072 20.7276 8.07929C20.2488 8.28786 19.7322 8.3955 19.21 8.3955C18.6877 8.3955 18.1711 8.28786 17.6923 8.07929C17.2136 7.87072 16.7829 7.5657 16.4273 7.18327C16.0716 6.80085 15.7986 6.34922 15.6253 5.85658C15.452 5.36394 15.3821 4.84087 15.42 4.32H11.58C11.58 4.81903 11.4817 5.31316 11.2907 5.7742C11.0998 6.23524 10.8199 6.65415 10.467 7.00701C10.1141 7.35987 9.69522 7.63978 9.23418 7.83075C8.77314 8.02171 8.27901 8.12 7.77998 8.12C7.28096 8.12 6.78682 8.02171 6.32579 7.83075C5.86475 7.63978 5.44584 7.35987 5.09298 7.00701C4.74012 6.65415 4.46021 6.23524 4.26924 5.7742C4.07827 5.31316 3.97998 4.81903 3.97998 4.32C3.47927 4.32 2.9835 4.41895 2.52115 4.61116C2.05881 4.80338 1.63901 5.08508 1.28588 5.44006C0.932758 5.79504 0.653269 6.21632 0.463481 6.67967C0.273694 7.14301 0.177349 7.6393 0.179984 8.14001V21.58C0.17997 22.5685 0.571258 23.5167 1.26832 24.2175C1.96539 24.9184 2.91154 25.3147 3.89998 25.32H14.86C14.777 24.7843 14.8226 24.2365 14.993 23.7218C15.1634 23.2072 15.4537 22.7404 15.84 22.36L17.56 20.62Z"
-      fill="#5D6C74"
-    />
-    <path
-      d="M18.84 6.54C19.3647 6.48617 19.8627 6.28203 20.2743 5.95209C20.6858 5.62215 20.9933 5.18045 21.16 4.68V2.82C20.9922 2.3184 20.6848 1.87507 20.2738 1.54207C19.8629 1.20906 19.3655 1.0002 18.84 0.940002C18.3484 0.950465 17.8804 1.15315 17.5364 1.50459C17.1925 1.85603 16.9999 2.32825 17 2.82V4.68C17 5.16986 17.1932 5.63995 17.5377 5.9882C17.8822 6.33645 18.3502 6.53473 18.84 6.54Z"
-      fill="#5D6C74"
-    />
-    <path
-      d="M26.18 20.62L26.76 20.04L29.44 17.38C29.564 17.2589 29.6625 17.1143 29.7298 16.9546C29.797 16.7948 29.8317 16.6233 29.8317 16.45C29.8317 16.2767 29.797 16.1052 29.7298 15.9455C29.6625 15.7857 29.564 15.6411 29.44 15.52L28.5 14.58C28.2593 14.369 27.9501 14.2527 27.63 14.2527C27.3099 14.2527 27.0007 14.369 26.76 14.58H26.64L23.4 17.84L20.14 14.58C20.0189 14.456 19.8743 14.3575 19.7146 14.2903C19.5549 14.223 19.3833 14.1884 19.21 14.1884C19.0367 14.1884 18.8652 14.223 18.7055 14.2903C18.5457 14.3575 18.4011 14.456 18.28 14.58L17.36 15.52C17.236 15.6411 17.1375 15.7857 17.0703 15.9455C17.003 16.1052 16.9684 16.2767 16.9684 16.45C16.9684 16.6233 17.003 16.7948 17.0703 16.9546C17.1375 17.1143 17.236 17.2589 17.36 17.38L20.6 20.62L17.36 23.88C17.1738 24.0639 17.0474 24.2998 16.9975 24.5567C16.9475 24.8137 16.9763 25.0797 17.08 25.32C17.1395 25.4798 17.2354 25.6236 17.36 25.74L18.28 26.66C18.4011 26.784 18.5457 26.8825 18.7055 26.9497C18.8652 27.017 19.0367 27.0517 19.21 27.0517C19.3833 27.0517 19.5549 27.017 19.7146 26.9497C19.8743 26.8825 20.0189 26.784 20.14 26.66L21.5 25.32L23.4 23.42L26.64 26.66C26.7611 26.784 26.9057 26.8825 27.0655 26.9497C27.2252 27.017 27.3967 27.0517 27.57 27.0517C27.7433 27.0517 27.9149 27.017 28.0746 26.9497C28.2343 26.8825 28.3789 26.784 28.5 26.66L29.44 25.74C29.564 25.6189 29.6625 25.4743 29.7298 25.3146C29.797 25.1548 29.8317 24.9833 29.8317 24.81C29.8317 24.6367 29.797 24.4652 29.7298 24.3055C29.6625 24.1457 29.564 24.0011 29.44 23.88L26.18 20.62Z"
-      fill="#5D6C74"
-    />
-  </svg>
-);
