--- src/components/TaskStatusIcon/WillRun.tsx
@@ -1,18 +1,0 @@
-export const WillRun: React.FC = () => (
-  <svg width="28" height="24" viewBox="0 0 28 24" fill="none">
-    <path
-      d="M6 2.55556C6 1.57372 6.89543 0.777779 8 0.777779C9.10457 0.777779 10 1.57372 10 2.55556V4.33333C10 5.31517 9.10457 6.11111 8 6.11111C6.89543 6.11111 6 5.31517 6 4.33333V2.55556Z"
-      fill="#21313C"
-    />
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M16 4.33333H12C12 6.29701 10.2091 7.88889 8 7.88889C5.79086 7.88889 4 6.29701 4 4.33333C1.79086 4.33333 0 5.92521 0 7.88889V20.3333C0 22.297 1.79086 23.8889 4 23.8889H24C26.2091 23.8889 28 22.297 28 20.3333V7.88889C28 5.92521 26.2091 4.33333 24 4.33333C24 6.29701 22.2091 7.88889 20 7.88889C17.7909 7.88889 16 6.29701 16 4.33333ZM22 11.4444H16V16.7778H22V11.4444Z"
-      fill="#21313C"
-    />
-    <path
-      d="M18 4.33333C18 5.31517 18.8954 6.11111 20 6.11111C21.1046 6.11111 22 5.31517 22 4.33333V2.55556C22 1.57372 21.1046 0.777779 20 0.777779C18.8954 0.777779 18 1.57372 18 2.55556V4.33333Z"
-      fill="#21313C"
-    />
-  </svg>
-);
